Refactored Login pages
