+++ my-app/src/App.jsx
@@ -1,140 +1,30 @@
 /* This is the home page where users will login to the room using a given roomCode*/
 
 import { Routes, Route, useNavigate, Navigate, useLocation } from 'react-router-dom';
-import Survey from "./survey/Survey";
-import './App.css';
-import { useState } from "react";
-import { loginUser, getCreatedUser } from '../services/usersService';
-import { getRoom, loginRoom } from '../services/roomsService';
-import { socket } from "./socket"
-import Interaction from "./components/interaction/Interaction";
-import Exit from "./Exit"
-import Admin from "./admin/Admin"
-import WaitingRoom from './components/waiting/WaitingRoom';
-import AdminWaitingRoom from './components/waiting/AdminWaitingRoom'
-import AdminInteraction from './components/interaction/AdminInteraction'
-import LoginAdmin from "./admin/AdminLogin";
+import "./App.css"; // might not need this later
+import { socket } from "./socket";
+import Login from "./components/login/Login";
+import WaitingRoom from "./components/waiting/WaitingRoom";
 import Instructions from './components/instructions/Instructions';
-import AdminInstructions from './components/instructions/AdminInstructions';
-import AdminSurvey from './admin/AdminSurvey';
+import Interaction from "./components/interaction/Interaction";
+import Survey from "./survey/Survey";
+import Exit from "./Exit";
+import AdminLogin from './components/login/AdminLogin';
+import Admin from "./admin/Admin";
 import CompletedRoomPage from './admin/CompletedRoomDetails';
+import AdminWaitingRoom from './components/waiting/AdminWaitingRoom';
+import AdminInstructions from './components/instructions/AdminInstructions';
+import AdminInteraction from './components/interaction/AdminInteraction';
+import AdminSurvey from './admin/AdminSurvey';
+
 
 function RequireState({ children, fallback = "/" }) {
-    const location = useLocation();
-    if (!location.state) {
-      return <Navigate to={fallback} replace />;
-    }
-    return children;
+  const location = useLocation();
+  if (!location.state) {
+    return <Navigate to={fallback} replace />;
   }
-
-function Home() {
-  const [name, setName] = useState("");
-  const navigate = useNavigate();
-  const [roomCode, setRoomCode] = useState("");
-  const [error, setError] = useState("");
-
-/** After the user hits submit it'll double check if it has a roomCode and a userName,
- *  then it will pass that information to the database and carry the userId to the next page
- */
-  async function handleClick(){
-    try {
-      setError("");
-      if (!name || !roomCode){
-        setError("Please enter your name and room code.");
-        return;
-      }
-
-      const parsedRoomCode = parseInt(roomCode);
-      if (isNaN(parsedRoomCode)) {
-        setError("Room code must be a number.");
-        return;
-      }
-
-      const canLogin = await loginRoom(parsedRoomCode);
-      if (!canLogin) {
-        setError("Invalid room code.");
-        return;
-      }
-
-      const room = await getRoom(parsedRoomCode);
-      const userIds = Array.isArray(room.userIds) ? room.userIds : JSON.parse(room.userIds);
-      if(room.completed) {
-        setError("This session has already ended.");
-        return;
-      }
-      if (userIds.length > 0) {
-        const user = await getCreatedUser(name, parsedRoomCode);
-        if (!user) {
-          setError("This room has already started. You are not part of this session.");
-          return;
-        }
-        
-        const status = room.status;
-        if (status === "waiting") {
-            navigate("/waiting", { state: { user }});
-        } else if (status === "instructions") {
-            navigate("/instructions", { state: { user }});
-        } else if (status === "interaction") {
-            navigate("/interaction", { state: { user }});
-        } else if (status === "survey") {
-            navigate("/survey", { state: { user }});
-        }
-        return;
-      }
-      const user = await loginUser(name, parsedRoomCode);
-      navigate("/waiting", { state: { user } }); 
-    } catch (err) {
-      setError(err.message || "Something went wrong. Please try again.");
-    }
-  };
-
-  /** This allows the enter button to be hit and send it to the function above to login */
-  const handleKeyDown = (e) => {
-    if (e.key === "Enter") {
-      e.preventDefault();
-      handleClick();
-    }
-  };
-
-  return(
-    <div className="login-page">
-      <div className="login-card">
-        <h1 className="login-title">Join your session</h1>
-        <p className="login-subtitle">Enter your details to join the room</p>
-        <form className="login-form" onSubmit={(e) => { e.preventDefault(); handleClick(); }}>
-          <div className="form-field">
-            <label htmlFor="name">Your name</label>
-            <input
-              id="name"
-              type="text"
-              value={name}
-              onChange={(e) => setName(e.target.value)}
-              onKeyDown={handleKeyDown}
-              placeholder="e.g. Alex Johnson"
-              autoComplete="name"
-            />
-          </div>
-          <div className="form-field">
-            <label htmlFor="roomCode">Room code</label>
-            <input
-              id="roomCode"
-              type="text"
-              value={roomCode}
-              onChange={(e) => setRoomCode(e.target.value)}
-              onKeyDown={handleKeyDown}
-              placeholder="e.g. 675435"
-            />
-          </div>
-          {error && <p className="form-error">{error}</p>}
-          <button type="submit" className="btn-primary" disabled={!name.trim() || !roomCode.trim()}>
-            Enter room
-          </button>
-        </form>
-      </div>
-    </div>
-  );
+  return children;
 }
-
 
 export default function App() {
   const location = useLocation();
@@ -142,7 +32,7 @@ export default function App() {
   const hideHomeOn = ["/interaction", "/waiting", "/survey", "/instructions"];
   const shouldHideHome = hideHomeOn.includes(location.pathname)
   const isAdminPage = location.pathname.startsWith("/admin");
-  const isAdminLogin = location.pathname === "/adminLogin";
+  const isAdminLogin = location.pathname === "/admin/login";
 
   function handleLogout() {
     sessionStorage.removeItem("admin");
@@ -175,20 +65,20 @@ export default function App() {
       </header>
 
       <Routes>
-        <Route path="/" element={<Home />} />
+        <Route path="/" element={<Login />} />
         <Route path="/survey" element={<RequireState> <Survey /> </RequireState>} /> 
         <Route path="/interaction" element={<RequireState> <Interaction /> </RequireState>} /> 
         <Route path="/exit" element={<RequireState> <Exit /> </RequireState>}/>
         <Route path="/admin" element={<Admin />}/>
         <Route path="/waiting" element={<RequireState> <WaitingRoom /> </RequireState>} />
-        <Route path="/admin/waiting" element={<RequireState fallback="/adminLogin"> <AdminWaitingRoom /></RequireState>} />
-        <Route path="/admin/interaction" element={<RequireState fallback="/adminLogin"> <AdminInteraction/> </RequireState>} />
-        <Route path='/adminLogin' element={<LoginAdmin/>} />
+        <Route path="/admin/waiting" element={<RequireState fallback="/admin/login"> <AdminWaitingRoom /></RequireState>} />
+        <Route path="/admin/interaction" element={<RequireState fallback="/admin/login"> <AdminInteraction/> </RequireState>} />
+        <Route path='/admin/login' element={<AdminLogin/>} />
         <Route path='/instructions' element={<RequireState> <Instructions/> </RequireState>} /> 
-        <Route path='/admin/instructions' element={<RequireState fallback="/adminLogin"><AdminInstructions/></RequireState>} />
-        <Route path='/admin/survey' element={<RequireState fallback="/adminLogin"> <AdminSurvey/> </RequireState>} />
+        <Route path='/admin/instructions' element={<RequireState fallback="/admin/login"><AdminInstructions/></RequireState>} />
+        <Route path='/admin/survey' element={<RequireState fallback="/admin/login"> <AdminSurvey/> </RequireState>} />
         {/* add a route to llm page when its added */}
-        <Route path="*" element={window.location.pathname.includes("admin") ? ( < Navigate to= "/adminLogin" /> ) : ( <Navigate to="/" />)} />
+        <Route path="*" element={window.location.pathname.includes("admin") ? ( < Navigate to= "/admin/login" /> ) : ( <Navigate to="/" />)} />
         <Route path="/admin/completed-room/:roomCode" element={<CompletedRoomPage />} />
       </Routes>
     </>
