+++ my-app/src/components/login/login.css
@@ -1,0 +1,218 @@
+/* === User Login (Professional) === */
+.login-page {
+  min-height: 100vh;
+  width: 100%;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  padding: 2rem;
+  box-sizing: border-box;
+}
+
+.login-card {
+  width: 100%;
+  max-width: 400px;
+  margin: 0 auto;
+  padding: 2.5rem;
+  text-align: center;
+  background: #ffffff;
+  border: 1px solid #e5e7eb;
+  border-radius: 12px;
+  box-shadow: 0 10px 25px -5px rgba(0, 0, 0, 0.08), 0 4px 6px -4px rgba(0, 0, 0, 0.04);
+  animation: fadeIn 0.4s ease forwards;
+}
+
+.login-title {
+  font-size: 1.5rem;
+  font-weight: 600;
+  color: #1e293b;
+  margin: 0 0 0.25rem 0;
+  letter-spacing: -0.02em;
+}
+
+.login-subtitle {
+  font-size: 0.875rem;
+  color: #64748b;
+  margin: 0 0 1.75rem 0;
+}
+
+.login-form {
+  width: 100%;
+  text-align: left;
+}
+
+.login-form .form-field {
+  margin-bottom: 1.25rem;
+}
+
+.login-form .form-field label {
+  display: block;
+  font-size: 0.875rem;
+  font-weight: 500;
+  color: #475569;
+  margin-bottom: 0.5rem;
+  text-align: left;
+}
+
+.login-form .form-field input {
+  width: 100%;
+  height: 44px;
+  padding: 0 0.875rem;
+  font-size: 1rem;
+  border: 1px solid #e2e8f0;
+  border-radius: 8px;
+  background: #ffffff;
+  max-width: none;
+  margin: 0;
+  box-shadow: none;
+  box-sizing: border-box;
+}
+
+.login-form .form-field input:focus {
+  outline: none;
+  border-color: #0ea5e9;
+  box-shadow: 0 0 0 3px rgba(14, 165, 233, 0.15);
+}
+
+.login-form .form-error {
+  font-size: 0.875rem;
+  color: #dc2626;
+  margin: 0 0 1rem 0;
+  padding: 0.5rem 0;
+}
+
+.login-form .btn-primary {
+  width: 100%;
+  height: 44px;
+  margin-top: 0.5rem;
+  font-size: 1rem;
+  font-weight: 500;
+  background: #0ea5e9;
+  color: #fff;
+  border: none;
+  border-radius: 8px;
+  cursor: pointer;
+  transition: all 0.2s ease;
+  text-align: center;
+}
+
+.login-form .btn-primary:hover:not(:disabled) {
+  background: #0284c7;
+}
+
+.login-form .btn-primary:disabled {
+  opacity: 0.5;
+  cursor: not-allowed;
+}
+
+/* === Admin Login (Dark Mode) === */
+.admin-container {
+  min-height: 100vh;
+  width: 100%;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  padding: 2rem;
+  box-sizing: border-box;
+  background: linear-gradient(180deg, #0f172a 0%, #1e293b 100%);
+}
+
+.admin-login-card {
+  width: 100%;
+  max-width: 400px;
+  margin: 0 auto;
+  padding: 2.5rem;
+  text-align: center;
+  background: rgba(30, 41, 59, 0.8);
+  border: 1px solid rgba(71, 85, 105, 0.4);
+  border-radius: 12px;
+  box-shadow: 0 10px 25px -5px rgba(0, 0, 0, 0.3), 0 4px 6px -4px rgba(0, 0, 0, 0.2);
+  animation: fadeIn 0.4s ease forwards;
+}
+
+.admin-login-header h1 {
+  font-size: 1.5rem;
+  font-weight: 600;
+  color: #e2e8f0;
+  margin: 0 0 0.25rem 0;
+  letter-spacing: -0.02em;
+}
+
+.admin-login-header p {
+  font-size: 0.875rem;
+  color: #94a3b8;
+  margin: 0 0 1.75rem 0;
+}
+
+.admin-login-card form {
+  width: 100%;
+  text-align: left;
+}
+
+.admin-login-card input {
+  width: 100%;
+  height: 44px;
+  padding: 0 0.875rem;
+  font-size: 1rem;
+  border: 1px solid rgba(71, 85, 105, 0.5);
+  border-radius: 8px;
+  background: rgba(15, 23, 42, 0.6);
+  color: #e2e8f0;
+  max-width: none;
+  margin: 0 0 0.75rem 0;
+  box-shadow: none;
+  box-sizing: border-box;
+}
+
+.admin-login-card input::placeholder {
+  color: #64748b;
+}
+
+.admin-login-card input:focus {
+  outline: none;
+  border-color: #0ea5e9;
+  box-shadow: 0 0 0 3px rgba(14, 165, 233, 0.2);
+}
+
+.admin-login-error {
+  font-size: 0.875rem;
+  color: #fca5a5;
+  margin: 0 0 1rem 0;
+  padding: 0.5rem 0;
+}
+
+.admin-login-submit {
+  width: 100%;
+  height: 44px;
+  margin-top: 0.5rem;
+  font-size: 1rem;
+  font-weight: 500;
+  background: #0ea5e9;
+  color: #fff;
+  border: none;
+  border-radius: 8px;
+  cursor: pointer;
+  transition: all 0.2s ease;
+  text-align: center;
+}
+
+.admin-login-submit:hover:not(:disabled) {
+  background: #0284c7;
+}
+
+.admin-login-submit:disabled {
+  opacity: 0.5;
+  cursor: not-allowed;
+}
+
+/* === Fade-in animation === */
+@keyframes fadeIn {
+  from {
+    opacity: 0;
+    transform: scale(0.97);
+  }
+  to {
+    opacity: 1;
+    transform: scale(1);
+  }
+}
